Generate different sitemaps for every part of the website
