+++ .gitignore
@@ -3,4 +3,4 @@
 node_modules
 public/blog.xml
 public/product-changelog.xml
-public/sitemap.xml
+public/sitemap*
